+++ metadata.rb
@@ -1,10 +1,9 @@
 name 'cpanel'
 maintainer 'Frank'
 maintainer_email 'escobarbfranklin@gmail.com'
-license 'The MIT License'
 description 'Installs the latest cPanel version.'
 long_description 'This Chef cookbook downloads and installs the latest cPanel software on the bootstrapped node.'
-version '0.1.1'
+version '0.1.2'
 %w( amazon centos redhat cloudlinux ).each do |os|
   supports os
 end