+++ recipes/default.rb
@@ -1,5 +1,5 @@
-#
 # Cookbook:: cpanel
 # Recipe:: default
 
-include_recipe 'cpanel::cpanel'
+include_recipe 'cpanel::install_cpanel'
+include_recipe 'cpanel::configure_cpanel'