+++ CHANGELOG.md
@@ -1,6 +1,10 @@
-# cpanel CHANGELOG
+# Chef cPanel Cookbook CHANGELOG
 
 This file is used to list changes made in each version of the cpanel cookbook.
+
+# 0.1.2
+Added attributes! You can now configure your cPanel/WHM installation.
+https://documentation.cpanel.net/display/70Docs/The+cpanel.config+File
 
 # 0.1.1
 Fixed issue with guard that's in place. Tested & working now! :)